Use relaxed semantics in CAS failures in loops
If a CAS in a loop fails, we do not need any synchronization with other
threads regarding other blocks in the bitmap, as we simply want to retry
setting the current block until we succeed (with sequentially consistent
semantics), or in some cases go on to another block, which will be loaded
with acquire semantics.

In other words, the load corresponding to a CAS failure has no ordering
requirements regarding reads or writes to different blocks.